requirements.txt
kaggle==1.5.1.1
google-cloud-storage==1.13.0